reorganize tests II
